refactor: move Handlebars template preloading logic to init module
